+++ Source/Core/TerrainEncoding.js
@@ -6,6 +6,7 @@ import defaultValue from "./defaultValue.js";
 import defined from "./defined.js";
 import CesiumMath from "./Math.js";
 import Matrix4 from "./Matrix4.js";
+import TerrainExaggeration from "./TerrainExaggeration.js";
 import TerrainQuantization from "./TerrainQuantization.js";
 
 var cartesian3Scratch = new Cartesian3();
@@ -23,25 +24,32 @@ var SHIFT_LEFT_12 = Math.pow(2.0, 12.0);
  * @alias TerrainEncoding
  * @constructor
  *
+ * @param {Cartesian3} center The center point of the vertices.
  * @param {AxisAlignedBoundingBox} axisAlignedBoundingBox The bounds of the tile in the east-north-up coordinates at the tiles center.
  * @param {Number} minimumHeight The minimum height.
  * @param {Number} maximumHeight The maximum height.
  * @param {Matrix4} fromENU The east-north-up to fixed frame matrix at the center of the terrain mesh.
  * @param {Boolean} hasVertexNormals If the mesh has vertex normals.
  * @param {Boolean} [hasWebMercatorT=false] true if the terrain data includes a Web Mercator texture coordinate; otherwise, false.
+ * @param {Boolean} [hasGeodeticSurfaceNormals=false] true if the terrain data includes geodetic surface normals; otherwise, false.
+ * @param {Number} [exaggeration=1.0] A scalar used to exaggerate terrain.
+ * @param {Number} [exaggerationRelativeHeight=0.0] The relative height from which terrain is exaggerated.
  *
  * @private
  */
 function TerrainEncoding(
+  center,
   axisAlignedBoundingBox,
   minimumHeight,
   maximumHeight,
   fromENU,
   hasVertexNormals,
-  hasWebMercatorT
+  hasWebMercatorT,
+  hasGeodeticSurfaceNormals,
+  exaggeration,
+  exaggerationRelativeHeight
 ) {
   var quantization = TerrainQuantization.NONE;
-  var center;
   var toENU;
   var matrix;
 
@@ -68,7 +76,6 @@ function TerrainEncoding(
       quantization = TerrainQuantization.NONE;
     }
 
-    center = axisAlignedBoundingBox.center;
     toENU = Matrix4.inverseTransformation(fromENU, new Matrix4());
 
     var translation = Cartesian3.negate(minimum, cartesian3Scratch);
@@ -119,7 +126,7 @@ function TerrainEncoding(
    * The center of the tile.
    * @type {Cartesian3}
    */
-  this.center = center;
+  this.center = Cartesian3.clone(center);
 
   /**
    * A matrix that takes a vertex from the tile, transforms it to east-north-up at the center and scales
@@ -151,6 +158,41 @@ function TerrainEncoding(
    * @type {Boolean}
    */
   this.hasWebMercatorT = defaultValue(hasWebMercatorT, false);
+
+  /**
+   * The terrain mesh contains geodetic surface normals, used for terrain exaggeration.
+   * @type {Boolean}
+   */
+  this.hasGeodeticSurfaceNormals = defaultValue(
+    hasGeodeticSurfaceNormals,
+    false
+  );
+
+  /**
+   * A scalar used to exaggerate terrain.
+   * @type {Number}
+   */
+  this.exaggeration = defaultValue(exaggeration, 1.0);
+
+  /**
+   * The relative height from which terrain is exaggerated.
+   */
+  this.exaggerationRelativeHeight = defaultValue(
+    exaggerationRelativeHeight,
+    0.0
+  );
+
+  /**
+   * The number of components in each vertex. This value can differ with different quantizations.
+   * @type {Number}
+   */
+  this.stride = 0;
+
+  this._offsetGeodeticSurfaceNormal = 0;
+  this._offsetVertexNormal = 0;
+
+  // Calculate the stride and offsets declared above
+  this._calculateStrideAndOffsets();
 }
 
 TerrainEncoding.prototype.encode = function (
@@ -160,7 +202,8 @@ TerrainEncoding.prototype.encode = function (
   uv,
   height,
   normalToPack,
-  webMercatorT
+  webMercatorT,
+  geodeticSurfaceNormal
 ) {
   var u = uv.x;
   var v = uv.y;
@@ -226,7 +269,73 @@ TerrainEncoding.prototype.encode = function (
     );
   }
 
+  if (this.hasGeodeticSurfaceNormals) {
+    vertexBuffer[bufferIndex++] = geodeticSurfaceNormal.x;
+    vertexBuffer[bufferIndex++] = geodeticSurfaceNormal.y;
+    vertexBuffer[bufferIndex++] = geodeticSurfaceNormal.z;
+  }
+
   return bufferIndex;
+};
+
+var scratchPosition = new Cartesian3();
+var scratchGeodeticSurfaceNormal = new Cartesian3();
+
+TerrainEncoding.prototype.addGeodeticSurfaceNormals = function (
+  oldBuffer,
+  newBuffer,
+  ellipsoid
+) {
+  if (this.hasGeodeticSurfaceNormals) {
+    return;
+  }
+
+  var oldStride = this.stride;
+  var vertexCount = oldBuffer.length / oldStride;
+  this.hasGeodeticSurfaceNormals = true;
+  this._calculateStrideAndOffsets();
+  var newStride = this.stride;
+
+  for (var index = 0; index < vertexCount; index++) {
+    for (var offset = 0; offset < oldStride; offset++) {
+      var oldIndex = index * oldStride + offset;
+      var newIndex = index * newStride + offset;
+      newBuffer[newIndex] = oldBuffer[oldIndex];
+    }
+    var position = this.decodePosition(newBuffer, index, scratchPosition);
+    var geodeticSurfaceNormal = ellipsoid.geodeticSurfaceNormal(
+      position,
+      scratchGeodeticSurfaceNormal
+    );
+
+    var bufferIndex = index * newStride + this._offsetGeodeticSurfaceNormal;
+    newBuffer[bufferIndex] = geodeticSurfaceNormal.x;
+    newBuffer[bufferIndex + 1] = geodeticSurfaceNormal.y;
+    newBuffer[bufferIndex + 2] = geodeticSurfaceNormal.z;
+  }
+};
+
+TerrainEncoding.prototype.removeGeodeticSurfaceNormals = function (
+  oldBuffer,
+  newBuffer
+) {
+  if (!this.hasGeodeticSurfaceNormals) {
+    return;
+  }
+
+  var oldStride = this.stride;
+  var vertexCount = oldBuffer.length / oldStride;
+  this.hasGeodeticSurfaceNormals = false;
+  this._calculateStrideAndOffsets();
+  var newStride = this.stride;
+
+  for (var index = 0; index < vertexCount; index++) {
+    for (var offset = 0; offset < newStride; offset++) {
+      var oldIndex = index * oldStride + offset;
+      var newIndex = index * newStride + offset;
+      newBuffer[newIndex] = oldBuffer[oldIndex];
+    }
+  }
 };
 
 TerrainEncoding.prototype.decodePosition = function (buffer, index, result) {
@@ -234,7 +343,7 @@ TerrainEncoding.prototype.decodePosition = function (buffer, index, result) {
     result = new Cartesian3();
   }
 
-  index *= this.getStride();
+  index *= this.stride;
 
   if (this.quantization === TerrainQuantization.BITS12) {
     var xy = AttributeCompression.decompressTextureCoordinates(
@@ -259,6 +368,39 @@ TerrainEncoding.prototype.decodePosition = function (buffer, index, result) {
   return Cartesian3.add(result, this.center, result);
 };
 
+TerrainEncoding.prototype.getExaggeratedPosition = function (
+  buffer,
+  index,
+  result
+) {
+  result = this.decodePosition(buffer, index, result);
+
+  var exaggeration = this.exaggeration;
+  var exaggerationRelativeHeight = this.exaggerationRelativeHeight;
+  var hasExaggeration = exaggeration !== 1.0;
+  if (hasExaggeration && this.hasGeodeticSurfaceNormals) {
+    var geodeticSurfaceNormal = this.decodeGeodeticSurfaceNormal(
+      buffer,
+      index,
+      scratchGeodeticSurfaceNormal
+    );
+    var rawHeight = this.decodeHeight(buffer, index);
+    var heightDifference =
+      TerrainExaggeration.getHeight(
+        rawHeight,
+        exaggeration,
+        exaggerationRelativeHeight
+      ) - rawHeight;
+
+    // some math is unrolled for better performance
+    result.x += geodeticSurfaceNormal.x * heightDifference;
+    result.y += geodeticSurfaceNormal.y * heightDifference;
+    result.z += geodeticSurfaceNormal.z * heightDifference;
+  }
+
+  return result;
+};
+
 TerrainEncoding.prototype.decodeTextureCoordinates = function (
   buffer,
   index,
@@ -268,7 +410,7 @@ TerrainEncoding.prototype.decodeTextureCoordinates = function (
     result = new Cartesian2();
   }
 
-  index *= this.getStride();
+  index *= this.stride;
 
   if (this.quantization === TerrainQuantization.BITS12) {
     return AttributeCompression.decompressTextureCoordinates(
@@ -281,7 +423,7 @@ TerrainEncoding.prototype.decodeTextureCoordinates = function (
 };
 
 TerrainEncoding.prototype.decodeHeight = function (buffer, index) {
-  index *= this.getStride();
+  index *= this.stride;
 
   if (this.quantization === TerrainQuantization.BITS12) {
     var zh = AttributeCompression.decompressTextureCoordinates(
@@ -297,7 +439,7 @@ TerrainEncoding.prototype.decodeHeight = function (buffer, index) {
 };
 
 TerrainEncoding.prototype.decodeWebMercatorT = function (buffer, index) {
-  index *= this.getStride();
+  index *= this.stride;
 
   if (this.quantization === TerrainQuantization.BITS12) {
     return AttributeCompression.decompressTextureCoordinates(
@@ -314,8 +456,7 @@ TerrainEncoding.prototype.getOctEncodedNormal = function (
   index,
   result
 ) {
-  var stride = this.getStride();
-  index = (index + 1) * stride - 1;
+  index = index * this.stride + this._offsetVertexNormal;
 
   var temp = buffer[index] / 256.0;
   var x = Math.floor(temp);
@@ -324,125 +465,124 @@ TerrainEncoding.prototype.getOctEncodedNormal = function (
   return Cartesian2.fromElements(x, y, result);
 };
 
-TerrainEncoding.prototype.getStride = function () {
-  var vertexStride;
+TerrainEncoding.prototype.decodeGeodeticSurfaceNormal = function (
+  buffer,
+  index,
+  result
+) {
+  index = index * this.stride + this._offsetGeodeticSurfaceNormal;
+
+  result.x = buffer[index];
+  result.y = buffer[index + 1];
+  result.z = buffer[index + 2];
+  return result;
+};
+
+TerrainEncoding.prototype._calculateStrideAndOffsets = function () {
+  var vertexStride = 0;
 
   switch (this.quantization) {
     case TerrainQuantization.BITS12:
-      vertexStride = 3;
+      vertexStride += 3;
       break;
     default:
-      vertexStride = 6;
+      vertexStride += 6;
   }
-
   if (this.hasWebMercatorT) {
-    ++vertexStride;
+    vertexStride += 1;
   }
-
   if (this.hasVertexNormals) {
-    ++vertexStride;
+    this._offsetVertexNormal = vertexStride;
+    vertexStride += 1;
+  }
+  if (this.hasGeodeticSurfaceNormals) {
+    this._offsetGeodeticSurfaceNormal = vertexStride;
+    vertexStride += 3;
   }
 
-  return vertexStride;
+  this.stride = vertexStride;
 };
 
-var attributesNone = {
+var attributesIndicesNone = {
   position3DAndHeight: 0,
   textureCoordAndEncodedNormals: 1,
+  geodeticSurfaceNormal: 2,
 };
-var attributes = {
+var attributesIndicesBits12 = {
   compressed0: 0,
   compressed1: 1,
+  geodeticSurfaceNormal: 2,
 };
 
 TerrainEncoding.prototype.getAttributes = function (buffer) {
   var datatype = ComponentDatatype.FLOAT;
   var sizeInBytes = ComponentDatatype.getSizeInBytes(datatype);
-  var stride;
+  var strideInBytes = this.stride * sizeInBytes;
+  var offsetInBytes = 0;
 
-  if (this.quantization === TerrainQuantization.NONE) {
-    var position3DAndHeightLength = 4;
-    var numTexCoordComponents = 2;
-
-    if (this.hasWebMercatorT) {
-      ++numTexCoordComponents;
-    }
-
-    if (this.hasVertexNormals) {
-      ++numTexCoordComponents;
-    }
-
-    stride = (position3DAndHeightLength + numTexCoordComponents) * sizeInBytes;
-
-    return [
-      {
-        index: attributesNone.position3DAndHeight,
-        vertexBuffer: buffer,
-        componentDatatype: datatype,
-        componentsPerAttribute: position3DAndHeightLength,
-        offsetInBytes: 0,
-        strideInBytes: stride,
-      },
-      {
-        index: attributesNone.textureCoordAndEncodedNormals,
-        vertexBuffer: buffer,
-        componentDatatype: datatype,
-        componentsPerAttribute: numTexCoordComponents,
-        offsetInBytes: position3DAndHeightLength * sizeInBytes,
-        strideInBytes: stride,
-      },
-    ];
-  }
-
-  var numCompressed0 = 3;
-  var numCompressed1 = 0;
-
-  if (this.hasWebMercatorT || this.hasVertexNormals) {
-    ++numCompressed0;
-  }
-
-  if (this.hasWebMercatorT && this.hasVertexNormals) {
-    ++numCompressed1;
-
-    stride = (numCompressed0 + numCompressed1) * sizeInBytes;
-
-    return [
-      {
-        index: attributes.compressed0,
-        vertexBuffer: buffer,
-        componentDatatype: datatype,
-        componentsPerAttribute: numCompressed0,
-        offsetInBytes: 0,
-        strideInBytes: stride,
-      },
-      {
-        index: attributes.compressed1,
-        vertexBuffer: buffer,
-        componentDatatype: datatype,
-        componentsPerAttribute: numCompressed1,
-        offsetInBytes: numCompressed0 * sizeInBytes,
-        strideInBytes: stride,
-      },
-    ];
-  }
-  return [
-    {
-      index: attributes.compressed0,
+  var attributes = [];
+  function addAttribute(index, componentsPerAttribute) {
+    attributes.push({
+      index: index,
       vertexBuffer: buffer,
       componentDatatype: datatype,
-      componentsPerAttribute: numCompressed0,
-    },
-  ];
+      componentsPerAttribute: componentsPerAttribute,
+      offsetInBytes: offsetInBytes,
+      strideInBytes: strideInBytes,
+    });
+    offsetInBytes += componentsPerAttribute * sizeInBytes;
+  }
+
+  if (this.quantization === TerrainQuantization.NONE) {
+    addAttribute(attributesIndicesNone.position3DAndHeight, 4);
+
+    var componentsTexCoordAndNormals = 2;
+    componentsTexCoordAndNormals += this.hasWebMercatorT ? 1 : 0;
+    componentsTexCoordAndNormals += this.hasVertexNormals ? 1 : 0;
+    addAttribute(
+      attributesIndicesNone.textureCoordAndEncodedNormals,
+      componentsTexCoordAndNormals
+    );
+
+    if (this.hasGeodeticSurfaceNormals) {
+      addAttribute(attributesIndicesNone.geodeticSurfaceNormal, 3);
+    }
+  } else {
+    // When there is no webMercatorT or vertex normals, the attribute only needs 3 components: x/y, z/h, u/v.
+    // WebMercatorT and vertex normals each take up one component, so if only one of them is present the first
+    // attribute gets a 4th component. If both are present, we need an additional attribute that has 1 component.
+    var usingAttribute0Component4 =
+      this.hasWebMercatorT || this.hasVertexNormals;
+    var usingAttribute1Component1 =
+      this.hasWebMercatorT && this.hasVertexNormals;
+    addAttribute(
+      attributesIndicesBits12.compressed0,
+      usingAttribute0Component4 ? 4 : 3
+    );
+
+    if (usingAttribute1Component1) {
+      addAttribute(attributesIndicesBits12.compressed1, 1);
+    }
+
+    if (this.hasGeodeticSurfaceNormals) {
+      addAttribute(attributesIndicesBits12.geodeticSurfaceNormal, 3);
+    }
+  }
+
+  return attributes;
 };
 
 TerrainEncoding.prototype.getAttributeLocations = function () {
   if (this.quantization === TerrainQuantization.NONE) {
-    return attributesNone;
+    return attributesIndicesNone;
   }
-  return attributes;
+  return attributesIndicesBits12;
 };
 
 TerrainEncoding.clone = function (encoding, result) {
+  if (!defined(encoding)) {
+    return undefined;
+  }
   if (!defined(result)) {
     result = new TerrainEncoding();
   }
@@ -456,6 +596,12 @@ TerrainEncoding.clone = function (encoding, result) {
   result.matrix = Matrix4.clone(encoding.matrix);
   result.hasVertexNormals = encoding.hasVertexNormals;
   result.hasWebMercatorT = encoding.hasWebMercatorT;
+  result.hasGeodeticSurfaceNormals = encoding.hasGeodeticSurfaceNormals;
+  result.exaggeration = encoding.exaggeration;
+  result.exaggerationRelativeHeight = encoding.exaggerationRelativeHeight;
+
+  result._calculateStrideAndOffsets();
+
   return result;
 };
 export default TerrainEncoding;