+++ Source/WorkersES6/createVerticesFromHeightmap.js
@@ -35,7 +35,7 @@ function createVerticesFromHeightmap(parameters, transferableObjects) {
 
   return {
     vertices: vertices.buffer,
-    numberOfAttributes: statistics.encoding.getStride(),
+    numberOfAttributes: statistics.encoding.stride,
     minimumHeight: statistics.minimumHeight,
     maximumHeight: statistics.maximumHeight,
     gridWidth: parameters.width,